+ Cleanup, destroy management
+ Plus minor refactor of DataView/DataViewModel
+ Plus minor refactor BaseFormField
