+++ cmp/form/index.js
@@ -6,6 +6,7 @@
  */
 export * from './Form';
 export * from './FormModel';
+export * from './field/BaseFieldModel';
 export * from './field/FieldModel';
 export * from './field/SubformsFieldModel';
 export * from './validation/ValidationState';
