+++ desktop/cmp/dataview/DataView.js
@@ -7,9 +7,8 @@
 
 import {Component} from 'react';
 import PT from 'prop-types';
-import {XH, HoistComponent, elemFactory, LayoutSupport} from '@xh/hoist/core';
-import {grid, GridModel} from '@xh/hoist/cmp/grid';
-import {omit} from 'lodash';
+import {HoistComponent, elemFactory, LayoutSupport} from '@xh/hoist/core';
+import {grid} from '@xh/hoist/cmp/grid';
 import {DataViewModel} from './DataViewModel';
 
 /**
@@ -29,49 +28,19 @@ export class DataView extends Component {
 
     baseClassName = 'xh-data-view';
 
-    constructor(props) {
-        super(props);
-        const {store, selModel, contextMenuFn, itemRenderer, emptyText} = props.model;
-        this._gridModel = new GridModel({
-            store,
-            selModel,
-            contextMenuFn,
-            emptyText,
-            columns: [
-                {
-                    colId: 'data',
-                    flex: true,
-                    elementRenderer: itemRenderer,
-                    agOptions: {
-                        valueGetter: this.valueGetter
-                    }
-                }
-            ]
-        });
-    }
-
     render() {
         const {rowCls, itemHeight, onRowDoubleClicked} = this.props;
         return grid({
             ...this.getLayoutProps(),
             className: this.getClassName(),
-            model: this._gridModel,
+            model: this.model.gridModel,
             agOptions: {
                 headerHeight: 0,
                 rowClass: rowCls,
                 getRowHeight: () => itemHeight
             },
-            onRowDoubleClicked: onRowDoubleClicked
+            onRowDoubleClicked
         });
-    }
-
-    valueGetter = (params) => {
-        const realData = omit(params.data.raw, 'id');
-        return Object.values(realData).join('\r');
-    };
-
-    destroy() {
-        XH.safeDestroy(this._gridModel);
     }
 }
